Updated Storyboard for the new fork and added IBOutlets and IBActions to code.
