+++ Cash Flow Tracker/Cash Flow Tracker WatchKit App/Assets.xcassets/AppIcon.appiconset/Contents.json
