+++ Cash Flow Tracker/Cash Flow Tracker WatchKit App/Info.plist
@@ -5,7 +5,7 @@
 	<key>CFBundleDevelopmentRegion</key>
 	<string>$(DEVELOPMENT_LANGUAGE)</string>
 	<key>CFBundleDisplayName</key>
-	<string>Daily Goal Tracker WatchKit App</string>
+	<string>$(PRODUCT_NAME)</string>
 	<key>CFBundleExecutable</key>
 	<string>$(EXECUTABLE_NAME)</string>
 	<key>CFBundleIdentifier</key>
@@ -26,7 +26,7 @@
 		<string>UIInterfaceOrientationPortraitUpsideDown</string>
 	</array>
 	<key>WKCompanionAppBundleIdentifier</key>
-	<string>joes.studio.Daily-Goal-Tracker</string>
+	<string>joes.studio.Cash-Flow-Tracker</string>
 	<key>WKWatchKitApp</key>
 	<true/>
 </dict>
