+++ Cash Flow Tracker/Cash Flow Tracker WatchKit Extension/Assets.xcassets/Complication.complicationset/Contents.json
