+++ Cash Flow Tracker/Cash Flow Tracker WatchKit Extension/Assets.xcassets/Complication.complicationset/Graphic Large Rectangular.imageset/Contents.json
